Single and Unordered Objective Nodes
Reworked the objective system to be based on nodes that can contain objectives in different forms but are still connected through a graph.  NodeSingle only holds one objective, while NodeUnordered can hold a list of objectives that can be completed in any order.  There is also a test scene for each type of node showing off how it works.