Node Ordered
Added a new node type that is similar to NodeUnordered but requries the objectives to be completed in the order of the list they are contained in.  Objectives that have canSkip as true can be skipped, but others cannot.  There is also a test scene illustrating the basic uses of the class.