Updated Objectives System
The objectives system has been significantly overhauled to make it more effective in the editor.  The main change is that actions are now assigned to objectives through the objectives themselves, which alleviates having to click around and find the action script every time you want to change it's objective or other properties.  The medicine and dosage also can be set a single time in the Scenario Groups rather than having to be set for every single action.  Finally, many of the scripts were changed from public to private using SerializeField, while others were hidden from the Inspector using HideInInspector.