Finished Room 2 Objectives
It still needs to be tested in VR, but it works for each medicine on keyboard.

I did notice a bug with the new placement system where triggers don't recognize exiting objects when they are picked up - it displays a warning message in the console that should help with debugging.