Basic Fill Minigame
The fill function currently works by increasing by a set amount for each click in non-VR mode and continuously increasing when touching the liquid bottle in VR mode (VR mode has not been tested yet).  This is not yet hooked up to the objective system and will need to be connected with the objective that tracks medicine dosage.